Navbar search working
