+++ TG-Netflix/src/mock-data/search_database.json
@@ -1,14 +1,102 @@
 [
-  {
-    "id": 1,
-    "title": "Batman"
-  },
-  {
-    "id": 2,
-    "title": "Superman"
-  },
-  {
-    "id": 3,
-    "title": "Watchmen"
-  }
+  { "id": 1, "title": "13th Letter, The" },
+  { "id": 2, "title": "Dreams (Kvinnodröm)" },
+  { "id": 3, "title": "Gasland Part II" },
+  { "id": 4, "title": "Trinity and Beyond" },
+  { "id": 5, "title": "Patsy" },
+  { "id": 6, "title": "Bad Lieutenant: Port of Call New Orleans" },
+  { "id": 7, "title": "She's All That" },
+  { "id": 8, "title": "Honeymoon" },
+  { "id": 9, "title": "Callas Forever" },
+  { "id": 10, "title": "Wood, The" },
+  { "id": 11, "title": "Last Command, The" },
+  { "id": 12, "title": "In Football We Trust" },
+  { "id": 13, "title": "Home of the Brave" },
+  { "id": 14, "title": "Silent Partner" },
+  { "id": 15, "title": "Journey to the West" },
+  { "id": 16, "title": "Gold Diggers: The Secret of Bear Mountain" },
+  { "id": 17, "title": "Things I Like, Things I Don't Like " },
+  { "id": 18, "title": "Christmas Party, The" },
+  { "id": 19, "title": "Twilight of the Ice Nymphs" },
+  { "id": 20, "title": "Leave It to Beaver" },
+  { "id": 21, "title": "Zeisters (Fat Guy Goes Nutzoid)" },
+  { "id": 22, "title": "Everything Must Go" },
+  { "id": 23, "title": "Contagion" },
+  { "id": 24, "title": "Rugrats Movie, The" },
+  { "id": 25, "title": "Tokyo Decadence (Topâzu)" },
+  { "id": 26, "title": "Battle for Haditha" },
+  { "id": 27, "title": "Actress, the Dollars and the Transylvanians" },
+  { "id": 28, "title": "Lady Liberty" },
+  { "id": 29, "title": "The Cyclone" },
+  { "id": 30, "title": "They Call Me Bruce" },
+  { "id": 31, "title": "Hierro " },
+  { "id": 32, "title": "You'll Find Out" },
+  { "id": 33, "title": "Whisper of Sin (Nuodemes uzkalbejimas)" },
+  { "id": 34, "title": "Killing Kasztner" },
+  { "id": 35, "title": "Mr. Sardonicus" },
+  { "id": 36, "title": "Inception" },
+  { "id": 37, "title": "Poison" },
+  { "id": 38, "title": "Je suis né d'une cigogne" },
+  { "id": 39, "title": "Testament of Orpheus, The (Testament d'Orphée)" },
+  { "id": 40, "title": "B.N.B. (Bunty Aur Babli)" },
+  { "id": 41, "title": "The Hire: Follow, The" },
+  { "id": 42, "title": "Terri" },
+  { "id": 43, "title": "On the Job" },
+  { "id": 44, "title": "Solyaris" },
+  { "id": 45, "title": "Susan Lenox (Her Fall and Rise)" },
+  { "id": 46, "title": "Collision: Christopher Hitchens vs. Douglas Wilson" },
+  { "id": 47, "title": "Dying Swan, The (Umirayushchii lebed)" },
+  { "id": 48, "title": "Goodbye, Columbus" },
+  { "id": 49, "title": "Crank" },
+  { "id": 50, "title": "Three Stooges in Orbit, The" },
+  { "id": 51, "title": "Mating of Millie, The" },
+  { "id": 52, "title": "There's Always Tomorrow" },
+  { "id": 53, "title": "Arthur and the Revenge of Maltazard" },
+  { "id": 54, "title": "7 Boxes (7 cajas)" },
+  { "id": 55, "title": "Final, The" },
+  { "id": 56, "title": "Vollidiot" },
+  { "id": 57, "title": "Great Gatsby, The" },
+  { "id": 58, "title": "Jönssonligans största kupp" },
+  { "id": 59, "title": "It! The Terror from Beyond Space" },
+  { "id": 60, "title": "Three Colors: Blue (Trois couleurs: Bleu)" },
+  { "id": 61, "title": "From Morn to Midnight (Von morgens bis Mitternacht)" },
+  { "id": 62, "title": "Born in East L.A." },
+  { "id": 63, "title": "360" },
+  { "id": 64, "title": "My Girlfriend's Boyfriend" },
+  { "id": 65, "title": "Mr. Warmth: The Don Rickles Project" },
+  { "id": 66, "title": "Made in Heaven" },
+  { "id": 67, "title": "L-Shaped Room, The" },
+  { "id": 68, "title": "Lucky Lady" },
+  { "id": 69, "title": "Jean de Florette" },
+  { "id": 70, "title": "The Snow Woman" },
+  { "id": 71, "title": "Who's the Caboose?" },
+  { "id": 72, "title": "Four Sided Triangle" },
+  { "id": 73, "title": "Palm Beach Story, The" },
+  { "id": 74, "title": "Mr. Robinson Crusoe" },
+  { "id": 75, "title": "Lift, De" },
+  { "id": 76, "title": "That's Life" },
+  { "id": 77, "title": "After the Life" },
+  { "id": 78, "title": "Good Time Girls, The (Bonnes femmes, Les)" },
+  { "id": 79, "title": "Antares" },
+  { "id": 80, "title": "Insidious: Chapter 2" },
+  { "id": 81, "title": "War on Democracy, The" },
+  { "id": 82, "title": "Seize the Day" },
+  { "id": 83, "title": "Fun Size" },
+  { "id": 84, "title": "Jab We Met" },
+  { "id": 85, "title": "Billy Bathgate" },
+  { "id": 86, "title": "High Roller: The Stu Ungar Story" },
+  { "id": 87, "title": "I'll Be There" },
+  { "id": 88, "title": "35 and Ticking" },
+  { "id": 89, "title": "Mission London" },
+  { "id": 90, "title": "Entre Amigos (Planta 4ª)" },
+  { "id": 91, "title": "My Man and I" },
+  { "id": 92, "title": "Wing Commander" },
+  { "id": 93, "title": "Being There" },
+  { "id": 94, "title": "Unholy" },
+  { "id": 95, "title": "Cord (Hide and Seek)" },
+  { "id": 96, "title": "We Live Again" },
+  { "id": 97, "title": "Real McCoy, The" },
+  { "id": 98, "title": "Marilyn in Manhattan" },
+  { "id": 99, "title": "Rat Race" },
+  { "id": 100, "title": "Hiroshima Mon Amour" }
 ]
