+++ TG-Netflix/src/pages/search/Search.jsx
@@ -1,41 +1,22 @@
 import Footer from '../../components/footer/footer';
 import * as S from './Search.styled';
 import Films from '../../mock-data/search_database.json';
-import { useSearchParams, useNavigate, useLocation } from 'react-router-dom';
-import { useEffect } from 'react';
+import { useSearchParams } from 'react-router-dom';
 
 export default function Search() {
-  const [search, setSearch] = useSearchParams();
-  const navigate = useNavigate();
-  const location = useLocation();
-  const query = search.get('q')
+  const [search] = useSearchParams();
+  const query = search.get('q');
 
-  const byTitle = (title) => (data) =>
-    data.title.toLowerCase().includes((title || '').toLowerCase());
-
-  const handleChange = (event) => {
-    setSearch({ q: event.target.value });
-  };
-
-  // useEffect(() => {
-  //   if (!search.get('q') && location.pathname === '/search') {
-  //     navigate('/browse');
-  //   }
-  // }, [search]);
+  const matchTitle = (query) => (data) =>
+    data.title.toLowerCase().includes((query || '').toLowerCase());
 
   return (
     <>
       <div className="padding-container">
         <S.Container>
-          <input
-            type="search"
-            value={query}
-            onChange={handleChange}
-          />
           <h3>Search results:</h3>
-
           <ul>
-            {Films.filter(byTitle(query)).map((data) => (
+            {Films.filter(matchTitle(query)).map((data) => (
               <li key={data.id}>{data.title}</li>
             ))}
           </ul>
